separate out tracker step from the old grade script
